pyproject.toml
[project]
name = "VOLI"
version = "0.1.0"
description = "MCP server for forex session volatility analysis"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "pytz>=2024.1",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
]

[project.scripts]
forex-session-mcp = "src.server:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"